NDLA-133 moved RankableConnectionUpdater to domain, split restart of docker out to separate cmd script so it can be used in run configurations
